+++ src/geometry/Polygon/RectAngle.js
@@ -6,25 +6,11 @@
 import * as maptalks from 'maptalks'
 const Coordinate = maptalks.Coordinate
 
-maptalks.Polygon.getCoordinateFromExtent = function (extent) {
-  let [minX, minY, maxX, maxY] = [extent[0], extent[1], extent[2], extent[3]]
-  return [minX, minY, minX, maxY, maxX, maxY, maxX, minY, minX, minY]
-}
-
-const boundingExtent = function (coordinates) {
-  let extent = new maptalks.Extent()
-  for (let i = 0, ii = coordinates.length; i < ii; ++i) {
-    extent.add(extent, coordinates[i])
-  }
-  return extent
-}
-
 class RectAngle extends maptalks.Polygon {
   constructor (coordinates, options = {}) {
     super(options)
     this.type = 'RectAngle'
     this._coordinates = []
-    this.isFill = ((options['isFill'] === false) ? options['isFill'] : true)
     if (coordinates) {
       this.setPoints(coordinates)
     }
@@ -39,18 +25,10 @@ class RectAngle extends maptalks.Polygon {
     let _points = Coordinate.toNumberArrays(this._coordinates)
     if (count < 2) return
     if (count === 2) {
-      let coordinates = []
-      if (this.isFill) {
-        let extent = boundingExtent(this._coordinates)
-        coordinates = maptalks.Polygon.getCoordinateFromExtent(extent)
-      } else {
-        let start = _points[0]
-        let end = _points[1]
-        coordinates = [start, [start[0], end[1]], end, [end[0], start[1]], start]
-      }
-      this.setCoordinates([
-        Coordinate.toCoordinates(coordinates)
-      ])
+      let start = _points[0]
+      let end = _points[1]
+      let coordinates = [start, [start[0], end[1]], end, [end[0], start[1]], start]
+      this.setCoordinates(Coordinate.toCoordinates(coordinates))
     }
   }
 
@@ -78,9 +56,9 @@ class RectAngle extends maptalks.Polygon {
 
   static fromJSON (json) {
     const feature = json['feature']
-    const attackArrow = new RectAngle(json['coordinates'], json['width'], json['height'], json['options'])
-    attackArrow.setProperties(feature['properties'])
-    return attackArrow
+    const reactAngle = new RectAngle(json['coordinates'], json['options'])
+    reactAngle.setProperties(feature['properties'])
+    return reactAngle
   }
 }