+++ CHANGELOG.md
@@ -1,7 +1,11 @@
+## 0.0.2 (暂时添加所有符号类型)
+
+* 添加部分点，圆，椭圆，8种箭头类型符号
+
 ## 0.0.1 (完善PlotDraw)
 
 * 完善PlotDraw
-* 添加部分geometry构造类（Ployline，Arror，Polygon）
+* 添加部分geometry构造类（Ployline，Arror，Polygon, Flag）
 
 ## 0.0.0 (搭建框架)